==> marine/crustaceans/crustaceans_farms/STEP2_crustacean_farm_locations.Rmd <==
---
title: "Location of crustacean Farms"
author: "Gage Clawson (NCEAS, UCSB, OHI)"
date: "12/14/2020"
output: html_document
editor_options: 
  chunk_output_type: console
---

This script standardizes location of crustacean farms to be lat/long point data.


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(raster)
library(tidyverse)
library(stringr)
library(rgdal)
library(PBSmapping)
library(maptools)
library(sf)
library(lwgeom)
library(data.table)
library(here)
library(mapview)
library(janitor)
library(rnaturalearth)
library(spatialEco)


path <- "/home/shares/food-systems/Food_footprint/_raw_data/Aquaculture_locations"

fao_crustacean_spp <- read_csv("marine/crustaceans/crustacean_farms/data/fao_mariculture_crustaceans_taxa_production.csv") %>%
  dplyr::filter(year == 2017)
fao_crustacean <- read_csv("marine/crustaceans/crustacean_farms/data/fao_mariculture_crustaceans.csv") %>%
  dplyr::filter(year == 2017)

unique(fao_crustacean$country)

```


# Raster

Standard raster used for this analysis 

```{r}

source(here::here("_spatial/template_raster.R"))

```

# Global regions

```{r}

gadm <- raster("/home/shares/food-systems/Food_footprint/_raw_data/Country_regions/gadm36mol_pyfill_10km.tif")

gadm_latlon <- projectRaster(gadm, food_raster)

gadm_codes <- read.csv("/home/shares/food-systems/Food_footprint/_raw_data/Country_regions/gadm_reg_codes.csv") %>% 
  mutate(reg_id = as.numeric(reg_id))

```

```{r}

gadm_df <- tabularaster::as_tibble(gadm_latlon, value = TRUE, cells = TRUE) %>% 
  mutate(cellvalue = round(cellvalue, 0))

gadm_reg <- dplyr::left_join(gadm_df, gadm_codes, by = c("cellvalue" = "reg_id"))

```


Now we will map the remaining countries with crustacean farms. Workflow will be as follows: 
 - Try to obtain exact locations of farms from published source or 
 dataset (such as FAO or a research paper)
 - If there are not exact locations or comprehensive locations, find the number of farms. 
 - If there is no information at all, note it. We will gapfill later. 
 
For each point, we will add a column for "data quality". It will be scored as very good (5), good (4), fair (3), bad (2), and very bad (1). 

5: Comprehensive number of exact farms 

4: Subset of exact locations, need to add a specific number of farms to existing locations. 
  - country|species|farm locations
  - country|species|additional points

3: A specific number of farms are known, but no locations 
 - country|species|additional points
 
2.5: A specific number of farms that are not comprehensive, so we will have to gapfill on top of those, if necessary. 
 - country, species, n_existing_points

2: Exact locations are a subset of total farms. But the number of total farms are unknown. Need to be gapfilled by tonnage. 

1: No information on number of farms at all. Need to be gapfilled by tonnage. 
 - country|species|# additional farms (this will be NA)
 
## Indonesia
Can't find any infomration on crab farm locations in Indonesia. This will be fully gapfilled. data_quality = 1

Source: 
NA 

```{r}
idn_gf <- data.frame(iso3c = "IDN", type = "crustaceans", num_additional_farms = NA, data_quality = 1)
write_csv(idn_gf, here("marine/crustaceans/crustacean_farms/data/no_info_1/gf_idn.csv")) 

```

## China 

Current data is from https://zenodo.org/record/3881612, only 93 farms, so probably not comprehensive, just a drop in the bucket compared to the produciton of China. There are likely many more farms than this. Because we can't find any information on more specific locations or specific number of farms, we will gapfill later based on global tonnes/farm. For now we need to mark the data quality as 2. 

```{r}
# chn_gf <- data.frame(iso3c = "CHN", type = "crustaceans", num_additional_farms = NA, data_quality = 1)
# write_csv(chn_gf, here("marine/crustaceans/crustacean_farms/data/no_info_1/gf_chn.csv")) 

farms_chn <- read.csv("/home/shares/food-systems/Food_footprint/_raw_data/Aquaculture_locations/China/split_spp/farms_crust_chn.csv") %>%
  dplyr::select(-id) %>%
  mutate(data_quality = 2)

write_csv(farms_chn, here("marine/crustaceans/crustaceans_farms/data/known_farm_sites_2_4_5/farms_chn.csv"))

```


## Bangladesh

https://www.researchgate.net/publication/318250461_Mud_crab_aquaculture_and_fisheries_in_coastal_Bangladesh

"Mud crabs have already gained popularity among coastal communities in greater Khulna and Chittagong regions (Azam et al. 1998) and are being recognized as a candidate species for culture in brackishwater environments (Ballao et al. 1999). There are about 300,000 people directly or indirectly connected with mud crab farming activities."

Source: 
Rahman, M., Islam, M.A., Haque, S.M., Wahab, A., 2017. Mud Crab Aquaculture and Fisheries in Coastal Bangladesh 7.

 - Even though we have this decent information, it is not location or number based. Therefore this will be fully gapfilled. data_quality = 1. 
 
```{r}
bgd_gf <- data.frame(iso3c = "BGD", type = "crustaceans", num_additional_farms = NA, data_quality = 1)
write_csv(bgd_gf, here("marine/crustaceans/crustacean_farms/data/no_info_1/gf_bgd.csv")) 
```

## Myanmar 

http://www.fao.org/3/ad497e/ad497e05.htm
"Mud crab farming is also important in Myanmar. Mud crab (Scylla serrata, and related species) culture in mangroves or tidal flats is practiced mainly in the Ayeyarwaddy delta area, in Rakhine and southern parts of Myanmar."

Source:
Myanmar aquaculture and inland fisheries, 2003. Food and Agriculture Organization of the United Nations Regional Office For Asia and the Pacific.

 - Even though we have this decent information, it is not location or number based. Therefore this will be fully gapfilled. data_quality = 1. 

```{r}
mmr_gf <- data.frame(iso3c = "MMR", type = "crustaceans", num_additional_farms = NA, data_quality = 1)
write_csv(mmr_gf, here("marine/crustaceans/crustacean_farms/data/no_info_1/gf_mmr.csv")) 
```


## Viet nam 

https://aciar.gov.au/sites/default/files/legacy/node/693/ias36_pdf_18085.pdf
"In addition to harvesting wild stock, mud crabs are now being farmed in
many coastal provinces in Vietnam, including Quang Ninh, Hai Phong,
Thanh Hoa, Ninh Binh, Nghe An, Thua Thien–Hue, Ba Ria–Vung Tau,
Ho Chi Minh City, Ben Tre, Tra Vinh, Soc Trang, Minh Hai, Kien Giang
and Bac Lieu."

"It is difficult to know how many people participate in this industry.
Apparently, no formal surveys have been conducted, there are no known
official statistics, and crab farming is typically undertaken together with
other activities, such as shrimp and/or fish aquaculture, other primary
production such as rice growing, or even as a sideline to paid employment."

"Table 4 gives the sole set of official statistics on crab aquaculture area and
production in Vietnam in 2004. The statistics cover both mud crabs and
blue swimmer crabs, and were collated by the Vietnamese Ministry of
Fisheries from reports of provincial fishery departments. "

 - Table on page 20: This is a table from 2004, but it will have to do. It also has production estimates, which is nice. 
 
Viet Nam: 
Lindner, B., 2005. IMPACTS OF MUD CRAB HATCHERY TECHNOLOGY IN VIETNAM 68.

 - Even though we have this decent information, it is not location or number based. Therefore this will be fully gapfilled. data_quality = 1. 

```{r}
vnm_gf <- data.frame(iso3c = "VNM", type = "crustaceans", num_additional_farms = NA, data_quality = 1)
write_csv(vnm_gf, here("marine/crustaceans/crustacean_farms/data/no_info_1/gf_vnm.csv")) 
```


## Philippines

All former reports can be found here: 
https://psa.gov.ph/content/fisheries-statistics-philippines

Fisheries statistics of the Philippines: https://psa.gov.ph/sites/default/files/Fisheries%20Statistics%20of%20the%20Philippines%2C%202017-2019.pdf

Table 125: TABLE 125 Volume of Mud Crab Production in Aquaculture by Region and Province,
 Philippines: 2017 - 2019
 

Consider using this data for future iterations... 
https://openstat.psa.gov.ph/PXWeb/pxweb/en/DB/search/?searchquery=aquaculture&rxid=bdf9d8da-96f1-4100-ae09-18cb3eaeb313

Source: 
Fisheries Statistics of the Philippines, 2017-2019, 2020. . Philippine Statistics Authority. [Accessed December 18, 2020]

 - Even though we have this decent information, it is not location or number based. Therefore this will be fully gapfilled. data_quality = 1. 


```{r}
phl_gf <- data.frame(iso3c = "PHL", type = "crustaceans", num_additional_farms = NA, data_quality = 1)
write_csv(phl_gf, here("marine/crustaceans/crustacean_farms/data/no_info_1/gf_phl.csv")) 
```
