==> marine/crustaceans/crustaceans_farms/STEP3_gapfill_crustaceans.Rmd <==
---
title: "Gapfilling the number of farms to add for crustacean countries based on tonnes of production"
author: "Gage Clawson (NCEAS, UCSB, OHI)"
date: "12/14/2020"
output: html_document
editor_options: 
  chunk_output_type: console
---


This script takes the number of farms information collected in STEP2 and calculates a global tonnes of production per farm, and calculates the number of remaining farms needed to place in our placement model. 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Read in all the relevant farm datasets... the only one that is represented for crustaceans are data_quality = 1.

```{r}

no_info_files <- list.files(here("marine/crustaceans/crustaceans_farms/data/no_info_1"), full=TRUE)
no_info_1 = plyr::ldply(no_info_files, read_csv)

farm_files <- list.files(here("marine/crustaceans/crustaceans_farms/data/known_farm_sites_2_4_5"), full=TRUE)
all_known_farms = plyr::ldply(farm_files, read_csv) %>%
  mutate(type = "crustaceans")

subset_gapfill_2 <- all_known_farms %>%
  filter(data_quality == 2)

```

Read in relevant production data
```{r}
crustaceans_tonnes <- read_csv(here::here("marine/crustaceans/crustaceans_farms/data/fao_mariculture_crustaceans.csv")) %>%
  mutate(type = "crustaceans") %>%
  filter(year == 2017)
```

Calculate a global tonnes per farm for crustaceans

 - We don't have any tonnes per farm estimates for crustaceans... Therefore we will use the shrimp tonnes/farm, as it is the most similar aquaculture practice to crustaceans.

```{r}
avg_global_shrimp_tonnes_per_farm = 336.5783
avg_global_shrimp_tonnes_per_farm # 336.5783
```

Now let's gapfill those that are data quality = 1
```{r}


## now lets do the locations that have no info and need gapfilling based on tonnes/farm
gapfill_1 <- no_info_1 %>%
  left_join(crustaceans_tonnes) %>%
  mutate(avg_tonnes_per_farm = avg_global_shrimp_tonnes_per_farm) %>%
  mutate(num_additional_farms = ceiling(fao_tonnes_production/avg_tonnes_per_farm)) %>%
  dplyr::select(iso3c, type, num_additional_farms, avg_tonnes_per_farm) %>%
  mutate(estimated = 1) %>%
  filter(num_additional_farms != 0)


```

Gapfill CHN (data quality = 2)

```{r}
## start with the subset of locations that need gapfilling based on tonnes/farm
gapfill_2 <- subset_gapfill_2 %>%
  group_by(iso3c, type) %>%
  summarise(n_existing_farms = n()) %>%
  ungroup() %>%
  left_join(crustaceans_tonnes) %>%
  mutate(avg_tonnes_per_farm = avg_global_shrimp_tonnes_per_farm) %>%
  mutate(tonnes_accounted_for = avg_tonnes_per_farm*n_existing_farms) %>%
  mutate(tonnes_remaining = fao_tonnes_production - tonnes_accounted_for) %>%
  mutate(tonnes_remaining = ifelse(tonnes_remaining <= 0, 0, tonnes_remaining)) %>%
  mutate(num_additional_farms = ceiling(tonnes_remaining/avg_tonnes_per_farm)) %>%
  dplyr::select(iso3c, type, num_additional_farms, avg_tonnes_per_farm) %>%
  mutate(estimated = 1) %>%
  filter(num_additional_farms != 0)


number_additional_points <- 
  rbind(gapfill_1, gapfill_2)
  
```


```{r}
  

## write this out to our data folder

write.csv(number_additional_points, here("marine/crustaceans/crustaceans_farms/output/crustaceans_points_to_add_final.csv"), row.names = FALSE)

all_known_farms_final <- all_known_farms %>%
  dplyr::select(-data_quality)


write.csv(all_known_farms_final, here("marine/crustaceans/crustaceans_farms/output/all_known_crustaceans_farms_final.csv"), row.names = FALSE)
```



Data check 

```{r}



test_2 <- number_additional_points %>%
  dplyr::select(iso3c)

length(unique(test_2$iso3c)) ## 6 countries


length(unique(crustaceans_tonnes$iso3c)) ## 6 countries


## Perfect

```
